feat: add pricing data to single fund pages
Refactor out the covalent price endpoint into a hook and use it on the single and combined fund pages to show latest price (if it has one).
